% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binner.R
\name{catBinner}
\alias{catBinner}
\title{Binning of categorical variable pairs}
\usage{
catBinner(x, y, dropPoints = FALSE)
}
\arguments{
\item{x}{factor vector for the first categorical variable}

\item{y}{factor vector for the second categorical variable}

\item{dropPoints}{logical; should points be dropped from final
bins?}
}
\value{
A list of lists each with elements `x`, `y`, `bnds`,
`expn`, `n`, and `stopped`.
}
\description{
`catBinner` converts the cross-tabulation of two
categorical variables into bins which work with all of the
functionality on bins built into `AssocBin`.
}
\details{
As both variables are already categorical, `catBinner`
performs no splits and does not merge any categories by default.
}
\author{
Chris Salahub
}
