% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{inDepPlot}
\alias{inDepPlot}
\title{Plot `inDep` output}
\usage{
inDepPlot(inDep, which = 1:5, ...)
}
\arguments{
\item{inDep}{output of an `inDep` function call}

\item{which}{indices of binnings to display ordered by increasing
p-value}
}
\value{
Nothing, but plot the specified indices.
}
\description{
Plot the specified indices of the output of an
`inDep` function call
}
\details{
For each index, this function plots a row of three plots.
The first plot is the raw data, the second plot is the ranks of
the data, and the final plot is the binning contained in the
output of `inDep`.
}
\author{
Chris Salahub
}
