% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopper.R
\name{makeCriteria}
\alias{makeCriteria}
\title{Make stop crteria}
\usage{
makeCriteria(...)
}
\arguments{
\item{...}{an arbitrary number of expressions which evaluate to
logicals}
}
\value{
A string which appends all expressions together.
}
\description{
Capture a sequence of logical statements and append
them into a single expression.
}
\details{
This function, along with `stopper` dictates the stop
behaviour of recursive binning. It accepts an arbitrary number
of arguments, each a logical statement, and appends them all into
a string separated by the pipe character.
}
\examples{
makeCriteria(depth >= 5, n < 1)
}
\author{
Chris Salahub
}
