% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{summary.inDep}
\alias{summary.inDep}
\alias{plot.inDep}
\title{S3 methods for `inDep`}
\usage{
\method{summary}{inDep}(inDep)

\method{plot}{inDep}(inDep, which = 1:5, ...)
}
\arguments{
\item{inDep}{object with class `inDep`}

\item{which}{indices of binnings to display in `inDep`, where
binnings are ordered by increasing p-value}

\item{...}{additional arguments to pass to `plot`}
}
\value{
Nothing for the plot method, while summary quietly returns
a summary of `inDep`
}
\description{
The `summary` and `plot` methods outlined here
support the quick description of an `inDep` object.
}
\details{
Methods

For each index in `which`, this function produces a row
of three plots. The first plot is the raw data, the second plot
is the ranks of the data, and the final plot is the binning
contained in the `inDep` object.
}
\section{Functions}{
\itemize{
\item \code{summary(inDep)}: Summary method for `genome`

\item \code{plot(inDep)}: Plot method for `genome`

}}
\author{
Chris Salahub
}
