% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{null10000}
\alias{null10000}
\title{Null replicates of different binning methods}
\format{
A list with four elements, the first gives a vector of the maximum
depths used, the follow three are arrays corresponding to
maximized chi, maximized mutual information, and random splitting.
The first dimension of each matrix gives a different statistic on
the final bins: the chi squared statistic, the mutual
information, the number of bins, and the maximum depth reached
before stopping. The second dimension corresponds to the maximum
depth from the first element of the list. The final dimension
provides the value of each statistic for each maximum depth over
10,000 replicates.
}
\usage{
data(null10000)
}
\description{
These data sets are statistic values for maximized chi, maximized
mutual information, and random binning applied to many replicates
of independent uniform pairs for a range of different maximum
depths. They can be used to determine empirical p-values of the
maximized methods. The number following 'null' corresponds to the
sample size of the independent uniform pairs, so 'null100' is the
data set for samples of 100 independent uniform pairs.
}
\keyword{datasets}
