% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{pairwiseAssociation}
\alias{pairwiseAssociation}
\title{Pairwise binning of complex data}
\usage{
pairwiseAssociation(
  data,
  stopCriteria,
  dualCategorical,
  categoricalContinuous,
  dualContinuous,
  measure = binChi,
  returnBinnings = FALSE
)
}
\arguments{
\item{data}{`data.frame` or object coercible to a `data.frame`}

\item{stopCriteria}{output of `makeCriteria` providing criteria
used to stop binning to be passed to binning functions}

\item{dualCategorical}{function which bins pairs of categorical
variables `x` and `y`}

\item{categoricalContinuous}{function which bins mixed variable
combinations and always assumes `x` is the categorical variable}

\item{dualContinuous}{function which bins pairs of continuous
variables `x` and `y`}

\item{measure}{function which measures association on the binnings
of all pairs}

\item{returnBinnings}{logical; should binnings be returned?}
}
\value{
A list of lists each with elements `x`, `y`, `bnds`,
`expn`, `n`, and `stopped`.
}
\description{
Measure association for every pair in a data set
by providing functions for each combination of data types.
}
\details{
This wrapper coordinates the application of recursive
binning to complex data containing potentially many types.
}
\author{
Chris Salahub
}
