% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp500pseudo}
\alias{sp500pseudo}
\title{De-Garched S&P 500 returns}
\format{
A matrix with 755 rows and 461 columns, the rows correspond to
dates between 2007 and 2009 and the columns correspond to the
different S&P 500 constituent stocks.
}
\usage{
data(sp500pseudo)
}
\description{
This data uses code from the 'zenplots' package to process S&P 500
consituent stock returns into uniform pseudo-observations for
measuring association.
}
\keyword{datasets}
