% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopper.R
\name{stopper}
\alias{stopper}
\title{Check bins against stop criteria}
\usage{
stopper(binList, criteria)
}
\arguments{
\item{binList}{a list of bins, each a list which can be cast as an
environment for evaluation}

\item{criteria}{string of logical expressions separated by pipes
to be evaluated within each bin of `binList`}
}
\value{
A logical vector of the same length as `binList`.
}
\description{
Evaluate the stop `criteria` for each bin in
`binList`
}
\details{
This function makes use of R's lexical scoping to
evaluate `criteria` (a string), within each bin of `binList`.
}
\author{
Chris Salahub
}
