% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{inDep}
\alias{inDep}
\title{Test pairwise variable independence}
\usage{
inDep(
  data,
  stopCriteria,
  catCon = uniRIntSplit,
  conCon = rIntSplit,
  dropPoints = FALSE
)
}
\arguments{
\item{data}{`data.frame` or object coercible to a `data.frame`}

\item{stopCriteria}{output of `makeCriteria` providing criteria
used to stop binning to be passed to binning functions}

\item{catCon}{splitting function to apply to pairs of one
cateogorical and one continuous variable}

\item{conCon}{splitting function to apply to pairs of continuous
variables}

\item{dropPoints}{logical; should returned bins contain points?}
}
\value{
A list of lists each with elements `x`, `y`, `bnds`,
`expn`, `n`, and `stopped`.
}
\description{
This is a high-level function which accepts a data
set, stop criteria, and split functions for continuous variables
and then applies a chi-square test for independence to bins
generated by recursively binning the ranks of continuous variables
or implied by the combinations of levels of categorical variables.
}
\details{
The output of `inDep` is a list, the first element of
which is a list of lists, each of which records the details of the
binning of a particular pair of variables
}
\author{
Chris Salahub
}
